controller/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-dotenv
psutil
Pillow
requests
pygame
RPi.GPIO
